+++ Unity-Technologies-networking/Editor/NetworkManagerEditor.cs
@@ -509,7 +509,7 @@ namespace UnityEditor
 
             if (EditorGUILayout.PropertyField(m_LogLevelProperty))
             {
-                LogFilter.currentLogLevel = (int)m_NetworkManager.logLevel;
+                LogFilter.currentLogLevel = m_NetworkManager.logLevel;
             }
 
             ShowScenes();
